src/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
pandas==2.1.3
pillow==10.1.0
pydantic==2.5.0
python-multipart==0.0.6
py-yaml=6.0.2
numpy==1.24.3
aiortc==1.6.0
aiofiles==23.2.1
python-socketio==5.10.0